requirements.txt
appwrite==3.0.0
langchain==0.0.352
langchain-google-genai==0.0.9
langchain-community
faiss-cpu==1.7.4
python-dotenv==1.0.0
